refactor: (#55) Main lunch 기능 UI 구조 분리
