+++ src/pages/main/ui/layout/MainTabSection.tsx
@@ -1,7 +1,7 @@
 import { Plus } from 'lucide-react';
 import type { MainTab } from '../../model/types';
 import MainBoardSection from '../MainBoardSection';
-import MainLunchMenuSection from '../MainLunchMenuSection';
+import MainLunchMenuSection from '../lunch/MainLunchMenuSection';
 import MainRankingSection from '../MainRankingSection';
 import { mockRooms } from '../room/mockRooms';
 import RoomCard from '../room/RoomCard';
